TravisCI down to GHC 7.8
Though I expect it to fail, I might as well be up-front about that.

Evidently the case is that 7.x boot files must have the same default
definitions as the non-boot files.  However, those defs do not (in 7.x)
have access to non-exported variables and boot files can not declare
non-exported values.